fix grape 0.11 with basic_auth bug.
older versions of grape did not use mustermann router. when using the
built-in http_basic middleware, unauthenticated requests' route is not
populated the environment within the route, and the route can't put out
useful information like route.route_path. In these cases we revert to
the older route info key in the rack env.